event/DeferEvent: add missing IsPending() check to Cancel()
Unlike boost::intrusive::list, our IntrusiveList class has no
is_linked() check.

+++ src/event/DeferEvent.hxx
@@ -77,7 +77,8 @@ public:
 	void ScheduleIdle() noexcept;
 
 	void Cancel() noexcept {
-		unlink();
+		if (IsPending())
+			unlink();
 	}
 
 protected:
